Properly setup the mixin in a module
